Dev/fastapi_gradio_practice/FastAPI/requirements.txt
fastapi==0.97
uvicorn
easyocr
python-multipart
pillow